Improve conversion of javadoc with html/taglets to asciidoc
This commit attempts to fix the generated asciidoc for javadoc blocks
that are not trivial:

A custom parsing of Roaster `getJavadoc()` model is used instead of
`getText()`/`getFullText()` to avoid mangling the original javadoc.
We attempt to convert a simple subset of HTML tags to their asciidoc
equivalents and to convert inline taglets to a relevant asciidoc
representation if any.

For HTML, we convert:
 - `p`, `br`, `b` and `i` to their direct equivalents
 - `strong` tags (including inline) to an `IMPORTANT:` admonition
 - `ul`/`ol` lists and their `li` elements in a best effort fashion

In case an `ol` is detected we have to turn ALL `li` to asciidoc
ordered list elements.

All other unknown HTML tags are removed but their node content is kept.

For taglets:
 - `@code` and `@value` taglets have their content turned into inline
  code blocks
 - `@link` and `@linkplain` taglets consider whether an alias text is
  provided. If so, only the alias text is provided in the asciidoc.
  If not, the target of the link is provided in the asciidoc as an
  inline code block.
 - unknown taglets are copied as an inline code block

Additionally, in order to ensure these asciidoc javadoc conversions are
correctly rendered in the output file, this commit polishes the syntax
of quotes and tables:
 - the `____` block style for quote is used instead of a single `>`
 - column format instruction `[cols="a,a"]` is used for tables

Fixes #21.