Include enclosing-class for comparison in [Metric|Span]Entry
Without this change, `[Metric|Span]Entry` only uses the enum name for
comparison.  When `[Meter|Span]Documentation` find the same name enums,
the list of entries gets overridden due to the comparison logic.
This change adds the comparison logic to also consider the
enclosing-class, so that the same enum value does not override the
existing entry.

Closes #99