Migrated to TOML versions (#129)
Replaces lock files with hard-coded versions managed in TOML. This is a prerequisite for having Dependabot upgrade dependencies for us with a Gradle build.

Maintain the behavior of when plugins are applied. It does not make a difference since there is only one project here. It will matter for other repos with multiple projects. This makes it easier to copy what was done here elsewhere.

Since we are no longer using lock files, we can remove tasks related to them.

---------

Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>