Upgrade dependencies with Dependabot (#146)
Configures dependabot to upgrade patch versions in maintenance branches and any version in `main`. We may want to fine tune this more, but this should be a start.

Whenever a new maintenance branch is created, it should be added to this config file. Likewise, whenever a branch is no longer maintained, it should be removed from this file so we no longer receive pull requests to it automatically.