Merged 1.0.x branch for plugins and non-plugins
I have no other idea how to make this work with dependabot. The difference is now that for 1.0.x branch we will be checking dependencies for plugins and upgrade them only by a patch, whereas before we would update the plugins by any value and we would check it for it weekly. Also previously we would use the gradle plugin portal to update only the com.gradle dependencies. With this change that constraint is removed